Fix remain compile error and refactor time-tag edit section relative things.
