+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Extends/RubyRomaji/RubyRomajiTagExtend.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using osu.Framework.Allocation;
+using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.RubyRomaji
@@ -11,13 +13,34 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.RubyRomaji
 
         public override float ExtendWidth => 350;
 
-        public TextTagExtend()
+        private Bindable<LyricEditorMode> bindableMode;
+
+        [BackgroundDependencyLoader]
+        private void load(ILyricEditorState state)
         {
-            Children = new Drawable[]
+            bindableMode = state.BindableMode.GetBoundCopy();
+            bindableMode.BindValueChanged(e =>
             {
-                new RubyTagEditSection(),
-                new RomajiTagEditSection(),
-            };
+                switch (e.NewValue)
+                {
+                    case LyricEditorMode.EditRuby:
+                        Children = new Drawable[]
+                        {
+                            new RubyTagEditSection(),
+                        };
+                        break;
+
+                    case LyricEditorMode.EditRomaji:
+                        Children = new Drawable[]
+                        {
+                            new RomajiTagEditSection(),
+                        };
+                        break;
+
+                    default:
+                        return;
+                }
+            }, true);
         }
     }
 }
