+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Components/RubyBlueprintContainer.cs
@@ -7,6 +7,7 @@ using osu.Framework.Bindables;
 using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Blueprints;
 using osu.Game.Rulesets.Karaoke.Objects;
+using osu.Game.Screens.Edit.Compose.Components;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components
 {
@@ -22,13 +23,26 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components
         }
 
         [BackgroundDependencyLoader]
-        private void load()
+        private void load(ILyricEditorState state)
         {
             // Add ruby tag into blueprint container
+            SelectedItems.BindTo(state.SelectedRubyTags);
             RegistBindable(rubyTags);
         }
 
+        protected override SelectionHandler<RubyTag> CreateSelectionHandler()
+            => new RubyTagSelectionHandler();
+
         protected override SelectionBlueprint<RubyTag> CreateBlueprintFor(RubyTag item)
             => new RubyTagSelectionBlueprint(item);
+
+        protected class RubyTagSelectionHandler : TextTagSelectionHandler
+        {
+            [BackgroundDependencyLoader]
+            private void load(ILyricEditorState state)
+            {
+                SelectedItems.BindTo(state.SelectedRubyTags);
+            }
+        }
     }
 }
